added hyperflow and amqp executor recipes

+++ deploy.sh
@@ -2,13 +2,19 @@
 
 # Usage: ./deploy.sh [host]
 
-host=$1
+if [ "$#" -ne 2 ]; then
+    echo "Usage: ./deploy.sh [attributes_file.json] [<user@>host]"
+    exit -1
+fi
+
+attributes_file=$1
+host=$2
 
 
-tar cj . | ssh -o 'StrictHostKeyChecking no' "$host" '
+tar cj . | ssh -o 'StrictHostKeyChecking no' "$host" "
 sudo rm -rf ~/chef &&
 mkdir ~/chef &&
 cd ~/chef &&
 tar xj &&
-sudo bash install.sh'
+sudo bash install.sh $attributes_file"
